+++ src/core/components/MidaMarketComponentOracle.ts
@@ -133,8 +133,8 @@ export class MidaMarketComponentOracle extends MidaTradingSystem {
         // </indicators>
 
         // <hooks>
-        await state.$component.periodUpdate?.call(state, period);
         await state.$component[`periodUpdate$${timeframe}`]?.call(state, period);
+        await state.$component.periodUpdate?.call(state, period);
         await state.$component.update?.call(state);
         await state.$component.lateUpdate?.call(state);
         // </hooks>
@@ -164,8 +164,8 @@ export class MidaMarketComponentOracle extends MidaTradingSystem {
         // </indicators>
 
         // <hooks>
-        await state.$component.periodClose?.call(state, period);
         await state.$component[`periodClose$${timeframe}`]?.call(state, period);
+        await state.$component.periodClose?.call(state, period);
         await state.$component.update?.call(state);
         await state.$component.lateUpdate?.call(state);
         // </hooks>
@@ -203,8 +203,7 @@ export class MidaMarketComponentOracle extends MidaTradingSystem {
                 indicatorInput = [ ...cappedPeriods.map((period) => period[priceKeys]), ];
             }
 
-            indicator.clear();
-            await indicator.next(indicatorInput);
+            await indicator.recalculate(indicatorInput);
         }));
     }
 }