+++ src/core/positions/MidaPosition.ts
@@ -114,7 +114,7 @@ export abstract class MidaPosition {
 
     public abstract getUnrealizedCommission (): Promise<MidaDecimal>;
 
-    public abstract changeProtection (protection: MidaProtectionDirectives): Promise<MidaProtectionChange>;
+    public abstract changeProtection (directives: MidaProtectionDirectives): Promise<MidaProtectionChange>;
 
     public abstract addVolume (volume: MidaDecimalConvertible): Promise<MidaOrder>;
 
@@ -185,6 +185,7 @@ export abstract class MidaPosition {
         }
 
         this.#emitter.notifyListeners("trade", { trade, });
+
         logger.info(`Position ${this.id} | Trade ${trade.id} executed`);
     }
 
@@ -200,26 +201,27 @@ export abstract class MidaPosition {
             trailingStopLoss: actualTrailingStopLoss,
         } = this.#protection;
 
-        if (takeProfit !== actualTakeProfit) {
+        if (takeProfit && !actualTakeProfit?.equals(takeProfit)) {
             this.#protection.takeProfit = takeProfit;
 
             this.#emitter.notifyListeners("take-profit-change", { takeProfit, });
         }
 
-        if (stopLoss !== actualStopLoss) {
+        if (stopLoss && !actualStopLoss?.equals(stopLoss)) {
             this.#protection.stopLoss = stopLoss;
 
             this.#emitter.notifyListeners("stop-loss-change", { stopLoss, });
         }
 
-        if (trailingStopLoss !== actualTrailingStopLoss) {
-            this.#protection.trailingStopLoss = actualTrailingStopLoss;
+        if (trailingStopLoss !== actualTrailingStopLoss && (stopLoss || actualStopLoss)) {
+            this.#protection.trailingStopLoss = trailingStopLoss;
 
             this.#emitter.notifyListeners("trailing-stop-loss-change", { trailingStopLoss, });
         }
 
         this.#emitter.notifyListeners("protection-change", { protection, });
-        logger.info(`Position ${this.id} | protection changed`);
+
+        logger.info(`Position ${this.id} | Protection changed`);
     }
 
     protected onSwap (swap: MidaDecimal): void {